feat: Search for people
Add support for /people/search. The form is much simpler than
/works/search, having only three fields. And a `User` object is much
simpler, needing only the username.

In theory, more of the user data could be parsed from the search
results, but calling user.reload() seems like a better choice.

I tried to keep the API consistent with AO3.Search, but chose
'all_fields' instead of 'any_field' in order to be consistent with the
AO3 people search form.

Update the version to 2.4.0 to indicate new capability added.